resolve svelte-fix errors (#106)
# Description
When running `npm run check` we should get no errors. Resolve all errors
or as many as possible.

- Closes #105 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)

# How to Test?

run `npm run check` and see that there is only the mouse exit/enter
errors. According to multiple forums, that is a known error that we
can't yet solve.

# Checklist:

- [ ] I have performed a self-review of my code
- [ ] I have commented my code & PR, particularly in hard-to-understand
areas
- [ ] I have checked at all the breakpoints to make sure it works on all
screen sizes

---------

Co-authored-by: Claire Olmstead <olmsteadclaire@gmail.com>